chore : set ts setting
ts 적용을 위합 파일 삭제
추가 라이브러리 설치
