--- src/config.js
@@ -1,6 +1,0 @@
-import { config } from "./env.js";
-
-export const DB_URL = config.DB_URL;
-export const PORT = config.PORT;
-export const ACCESS_TOKEN_SECRET = config.ACCESS_TOKEN_SECRET;
-export const REFRESH_TOKEN_SECRET = config.REFRESH_TOKEN_SECRET;
