--- src/constants/post.js
@@ -1,13 +1,0 @@
-const MIN_LENGTH_TITLE = 1;
-const MAX_LENGTH_TITLE = 50;
-const MIN_LENGTH_CONTENT = 1;
-const MAX_LENGTH_CONTENT = 1024;
-
-export const postSchema = {
-  MIN_LENGTH_TITLE,
-  MAX_LENGTH_TITLE,
-  MIN_LENGTH_CONTENT,
-  MAX_LENGTH_CONTENT,
-};
-
-export const DEFAULT_POST_PAGE_SIZE = 10;
