--- src/constants/product.js
@@ -1,25 +1,0 @@
-const MIN_LENGTH_NAME = 1;
-const MAX_LENGTH_NAME = 50;
-const MIN_LENGTH_DESCRIPTION = 1;
-const MAX_LENGTH_DESCRIPTION = 1024;
-const MIN_PRICE = 1;
-const MAX_PRICE = 9999999999;
-const MIN_COUNT_IMAGE = 0;
-const MAX_COUNT_IMAGE = 3;
-const MIN_COUNT_TAG = 0;
-const MAX_COUNT_TAG = 5;
-
-export const productSchema = {
-  MIN_LENGTH_NAME,
-  MAX_LENGTH_NAME,
-  MIN_LENGTH_DESCRIPTION,
-  MAX_LENGTH_DESCRIPTION,
-  MIN_PRICE,
-  MAX_PRICE,
-  MIN_COUNT_IMAGE,
-  MAX_COUNT_IMAGE,
-  MIN_COUNT_TAG,
-  MAX_COUNT_TAG,
-};
-
-export const DEFAULT_PRODUCT_PAGE_SIZE = 10;
