--- src/constants/tag.js
@@ -1,7 +1,0 @@
-const MIN_LENGTH_TAG = 1;
-const MAX_LENGTH_TAG = 50;
-
-export const tagSchema = {
-  MIN_LENGTH_TAG,
-  MAX_LENGTH_TAG,
-};
