--- src/constants/token.js
@@ -1,5 +1,0 @@
-// 임시로 테스트용 검증 기간 적용 export const ACCESS_TOKEN_EXPIRY = "10m";
-export const ACCESS_TOKEN_EXPIRY = "3d";
-export const REFRESH_TOKEN_EXPIRY = "7d";
-
-export const SALT_ROUNDS = 10;
