--- src/constants/user.js
@@ -1,23 +1,0 @@
-const MIN_LENGTH_NAME = 1;
-const MAX_LENGTH_NAME = 255;
-const MIN_LENGTH_NICKNAME = 1;
-const MAX_LENGTH_NICKNAME = 50;
-const MIN_LENGTH_PASSWORD = 8;
-const MAX_LENGTH_PASSWORD = 24;
-const MIN_LENGTH_EMAIL = 1;
-const MAX_LENGTH_EMAIL = 255;
-const MIN_LENGTH_IMAGE = 1;
-const MAX_LENGTH_IMAGE = 1024;
-
-export const userSchema = {
-  MIN_LENGTH_NAME,
-  MAX_LENGTH_NAME,
-  MIN_LENGTH_NICKNAME,
-  MAX_LENGTH_NICKNAME,
-  MIN_LENGTH_PASSWORD,
-  MAX_LENGTH_PASSWORD,
-  MIN_LENGTH_EMAIL,
-  MAX_LENGTH_EMAIL,
-  MIN_LENGTH_IMAGE,
-  MAX_LENGTH_IMAGE,
-};
