--- src/responses/comment-res.js
@@ -1,12 +1,0 @@
-export const commentSelect = {
-  id: true,
-  content: true,
-  createdAt: true,
-  User: {
-    select: {
-      id: true,
-      nickname: true,
-      image: true,
-    },
-  },
-};
