--- src/responses/user-res.js
@@ -1,8 +1,0 @@
-export const userSelect = {
-  id: true,
-  email: true,
-  name: true,
-  nickname: true,
-  image: true,
-  createdAt: true,
-};
