--- src/routes/article-comment.js
@@ -1,9 +1,0 @@
-/** POST */
-
-/** GET */
-
-/** GET */
-
-/** PATCH */
-
-/** DELETE */
