--- src/utils/error.js
@@ -1,5 +1,0 @@
-export function createCustomError({ status, message }) {
-  const err = new Error(message);
-  err.status = status;
-  return err;
-}
