--- src/utils/token.js
@@ -1,18 +1,0 @@
-import jwt from "jsonwebtoken";
-import {
-  ACCESS_TOKEN_EXPIRY,
-  REFRESH_TOKEN_EXPIRY,
-} from "../constants/token.js";
-import { ACCESS_TOKEN_SECRET, REFRESH_TOKEN_SECRET } from "../config.js";
-
-export function createAccessToken(userId) {
-  return jwt.sign({ id: userId }, ACCESS_TOKEN_SECRET, {
-    expiresIn: ACCESS_TOKEN_EXPIRY,
-  });
-}
-
-export function createRefreshToken(userId) {
-  return jwt.sign({ id: userId }, REFRESH_TOKEN_SECRET, {
-    expiresIn: REFRESH_TOKEN_EXPIRY,
-  });
-}
